easysync tests: Split into multiple files
